Revert "Final"

+++ src/main/java/br/com/fatesg/eventos/persistence/CategoriaPersistence.java
@@ -2,6 +2,6 @@ package br.com.fatesg.eventos.persistence;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import br.com.fatesg.eventos.model.Categoria;
+import br.com.fatesg.eventos.entities.Categoria;
 
-public interface CategoriaPersistence extends JpaRepository<Categoria, Integer> {}
+public interface CategoriaPersistence extends JpaRepository<Categoria, Long> {}
